requirements.txt
pyyaml
distilabel[hf-transformers, openai]>=1.0.0
datasets
transformers